refact:change the folder, endure cases
